kernel: simplify serial log flushing on panic
This adds an in_panic() function under panic.hpp which we use in the
serial driver. When not panicking, the serial driver acts as normal and
enables interrupts again and waits until it can write to it's internal
buffer. If we are panicking, transmit enough bytes to fill the buffer